Split tests for `CmdCtxBuilder` implementations to a module per scope.
